Enable using FreeBSD's base system libedit
On FreeBSD, libedit exists as part of the base system.  pkg-config(8)
and similar tools do not find the base system installation but will only
discover libedit if it exists on the system as part of a separately
installed package.

Listing PkgConfig before CBuilder in the alienfile prefers the package
if available, but falls back to the base system's version if present,
instead of installing from source.

The README.md changes in this commit appear to have taken place
automatically as part of the Dist::Zilla process: I did not make them
myself.